Update ES translation for Scroll Up module!
It has been very useful for me, everything that ArchBlood and GreenMeteor have delivered in my experience in humhub, without you it would not be possible to carry out my educational project. so I will be supporting the translation of the modules into Spanish since it is my native language.

Best regards.